Updated luigi routines ready for multiple developers
