+++ nesta/production/routines/health_data/nih_data/nih_collect_task.py
@@ -33,7 +33,7 @@ def exists(_class, **kwargs):
     return sql_exists().where(and_(*statements))
 
 
-class LatestDataToMySQL(luigi.Task):
+class CollectTask(luigi.Task):
     '''Scrape CSVs from the World ExPORTER site and dump the
     data in the MySQL server.
 
@@ -84,26 +84,3 @@ class LatestDataToMySQL(luigi.Task):
 
         # Mark the task as done
         self.output().touch()
-
-
-class RootTask(luigi.WrapperTask):
-    '''A dummy root task, which collects the database configurations
-    and executes the central task.
-
-    Args:
-        date (datetime): Date used to label the outputs
-        db_config_path (str): Path to the MySQL database configuration
-        production (bool): Flag indicating whether running in testing 
-                           mode (False, default), or production mode (True).
-    '''
-    date = luigi.DateParameter(default=datetime.date.today())
-    db_config_path = luigi.Parameter()
-    production = luigi.BoolParameter(default=False)
-
-    def requires(self):
-        '''Collects the database configurations
-        and executes the central task.'''
-        logging.getLogger().setLevel(logging.INFO)
-        yield LatestDataToMySQL(date=self.date, 
-                                db_config_path=self.db_config_path,
-                                production=self.production)
