+++ nesta/production/routines/health_data/nih_data/nih_root_task.py
@@ -1,0 +1,37 @@
+'''
+NIH data collection and processing
+==================================
+
+Luigi routine to collect NIH World RePORTER data
+via the World ExPORTER data dump. The routine
+transfers the data into the MySQL database before
+processing and indexing the data to ElasticSearch.
+'''
+
+import luigi
+import datetime
+import logging
+
+from nih_process_task import ProcessTask
+
+class RootTask(luigi.WrapperTask):
+    '''A dummy root task, which collects the database configurations
+    and executes the central task.
+
+    Args:
+        date (datetime): Date used to label the outputs
+        db_config_path (str): Path to the MySQL database configuration
+        production (bool): Flag indicating whether running in testing 
+                           mode (False, default), or production mode (True).
+    '''
+    date = luigi.DateParameter(default=datetime.date.today())
+    db_config_path = luigi.Parameter(default="mysqldb.config")
+    production = luigi.BoolParameter(default=False)
+
+    def requires(self):
+        '''Collects the database configurations
+        and executes the central task.'''
+        logging.getLogger().setLevel(logging.INFO)
+        yield ProcessTask(date=self.date,
+                          db_config_path=self.db_config_path,
+                          production=self.production)
